PR-001-ac-Implement readsensordata class

+++ WildBlueberrySensorSystem/src/Sensors/Potentiometer/Potentiometer_Analog.cpp
@@ -1,5 +1,5 @@
 #include "Potentiometer_Analog.h"
-
+// constructor
 Potentiometer_Analog::Potentiometer_Analog(int pin) : _data_pin(pin)
 {
     pinMode(pin, INPUT);
@@ -10,7 +10,7 @@ Potentiometer_Analog::~Potentiometer_Analog() {}
 
 float Potentiometer_Analog::readData()
 {
-    return static_cast<float>(analogRead(_data_pin));
+    return analogRead(_data_pin);
 }
 
 int Potentiometer_Analog::getPin()
